create base fixture
