+++ plugins/hal-view/tests/Fixture/ActorsFixture.php
@@ -3,36 +3,11 @@ declare(strict_types=1);
 
 namespace MixerApi\HalView\Test\Fixture;
 
-use Cake\TestSuite\Fixture\TestFixture;
-
 /**
  * ActorsFixture
  */
-class ActorsFixture extends TestFixture
+class ActorsFixture extends \App\Test\Fixture\ActorsFixture
 {
-    /**
-     * Fields
-     *
-     * @var array
-     */
-    // phpcs:disable
-    public $fields = [
-        'id' => ['type' => 'integer', 'length' => null, 'unsigned' => true, 'null' => false, 'default' => null, 'comment' => '', 'autoIncrement' => true, 'precision' => null],
-        'first_name' => ['type' => 'string', 'length' => 45, 'null' => false, 'default' => null, 'collate' => 'latin1_swedish_ci', 'comment' => '', 'precision' => null],
-        'last_name' => ['type' => 'string', 'length' => 45, 'null' => false, 'default' => null, 'collate' => 'latin1_swedish_ci', 'comment' => '', 'precision' => null],
-        'modified' => ['type' => 'datetime', 'length' => null, 'precision' => null, 'null' => false, 'default' => null, 'comment' => ''],
-        '_indexes' => [
-            'last_name' => ['type' => 'index', 'columns' => ['last_name'], 'length' => []],
-        ],
-        '_constraints' => [
-            'primary' => ['type' => 'primary', 'columns' => ['id'], 'length' => []],
-        ],
-        '_options' => [
-            'engine' => 'InnoDB',
-            'collation' => 'latin1_swedish_ci'
-        ],
-    ];
-    // phpcs:enable
     /**
      * Init method
      *
